# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для генерации персонализированных поздравлений: AI-изображения + песни. Аутентификация через httpOnly cookie (access_token).",
        "title": "FunBaza API",
        "contact": {
            "name": "FunBaza Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/sessions": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает список диалоговых сессий (тредов) текущего пользователя, отсортированных по дате обновления.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Список сессий",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Лимит (макс. 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SessionListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает сессию и все генерации в ней (тред). URL изображений и аудио — рабочие ссылки для скачивания.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Получить сессию с генерациями",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID сессии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SessionThreadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Изменяет заголовок (название) сессии. Макс. 300 символов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Переименовать сессию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID сессии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый заголовок",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.OkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "message": {
                    "type": "string",
                    "example": "Refresh token missing"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.ErrorDetail"
                }
            }
        },
...
        "handlers.OkResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handlers.SessionListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 30
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GenerationSession"
                    }
                }
            }
        },
        "handlers.SessionThreadResponse": {
            "type": "object",
            "properties": {
                "generations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GenerationRequest"
                    }
                },
                "session": {
                    "$ref": "#/definitions/models.GenerationSession"
                }
            }
        },
...
        "handlers.UpdateTitleRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 300,
                    "example": "Поздравление маме"
                }
            }
        },
...
        "models.GenerationRequest": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "credits_spent": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_count": {
                    "type": "integer"
                },
                "image_model": {
                    "type": "string"
                },
                "image_prompt": {
                    "type": "string"
                },
                "input_audio_key": {
                    "type": "string"
                },
                "input_audio_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "input_photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "parent_id": {
                    "type": "string"
                },
                "result_audios": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "result_images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_id": {
                    "type": "string"
                },
                "song_count": {
                    "type": "integer"
                },
                "song_lyrics": {
                    "type": "string"
                },
                "song_prompt": {
                    "type": "string"
                },
                "song_style": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.GenerationStatus"
                },
                "tariff_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.GenerationSession": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.GenerationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing_images",
                "processing_audio",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessingImages",
                "StatusProcessingAudio",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}